+++ app/[locale]/(tabs)/assets/page.tsx
@@ -1,11 +1,11 @@
-import { queryCustomers } from "../../components/chain_utils";
+import { queryCustomers } from "../../../../components/chain_utils";
 import AssetPanel from "./panel";
 
 export default async function Page({ searchParams }: {
     searchParams?: { [key: string]: string | string[] | undefined };
   }) {
     let page: number = 0;
-    if (searchParams.page){
+    if (searchParams && searchParams.page){
       let pageString = searchParams.page as string;
       let index = Number.parseInt(pageString);
       if (!Number.isNaN(index)){