+++ app/[locale]/(tabs)/layout.tsx
@@ -1,14 +1,14 @@
-import Navbar from "../../components/navbar";
-import SideTab from "../../components/side_tab";
+import Navbar from "../../../components/navbar";
+import SideTab from "../../../components/side_tab";
 
-export default function RequestLayout({ children }){
+export default function AssetLayout({ children }){
     return (
         <div>
             <Navbar/>
             <div className="container">
                 <div className="row">
                     <div className="col-2">
-                        <SideTab current='requests'/>
+                        <SideTab/>
                     </div>
                     <div className="col pt-3">
                         {children}