+++ app/[locale]/(tabs)/requests/detail/[id]/detail.tsx
@@ -1,7 +1,6 @@
 'use client';
-import { Verify } from 'crypto';
-import { useAppContext, getCurrentyFormatter } from '../../../../components/context';
-import { RequestStatus, VerifyMode, VerifyRequest } from '../../../../components/verify_request';
+import { useAppContext, getCurrentyFormatter } from '../../../../../../components/context';
+import { RequestStatus, VerifyMode, VerifyRequest } from '../../../../../../components/verify_request';
 
 const i18n = {
     en: {
@@ -83,7 +82,7 @@ export default function RequestDetail({data}:{
         },
         {
             label: texts.invokeTime,
-            value: new Date(invoke_time).toLocaleString(),
+            value: new Date(invoke_time as string).toLocaleString(),
         },
     ]
     let statusLabel: string, resultLabel: string;
@@ -117,7 +116,7 @@ export default function RequestDetail({data}:{
             },
             {
                 label: texts.completeTime,
-                value: new Date(verify_time).toLocaleString(),
+                value: new Date(verify_time as string).toLocaleString(),
             },
             {
                 label: texts.comment,