+++ app/layout.tsx
@@ -1,13 +1,8 @@
 import ChainProvider from '../components/chain_provider';
-import { ChainConnector } from '../components/chain_sdk';
-import { ContextData, ContextProvider } from '../components/context';
+import { ChainConnector } from "@taiyi-io/chain-connector";
 import { AccountStatus, AssetProperties, ASSET_SCHEMA_NAME } from '../components/customer_asset';
 import Boot from './bootstrap';
 import npmPackage from '../package.json';
-
-export const dynamic = 'force-dynamic',
-  revalidate = 0,
-  fetchCache = 'force-no-store';
 
 async function ensureSchema(conn: ChainConnector) {
   let hasSchema = await conn.hasSchema(ASSET_SCHEMA_NAME);
@@ -15,9 +10,8 @@ async function ensureSchema(conn: ChainConnector) {
     let defines = AssetProperties;
     await conn.createSchema(ASSET_SCHEMA_NAME, defines);
     console.log('schema %s initialized', ASSET_SCHEMA_NAME);
-      
   }
-  if (!await conn.hasDocument(ASSET_SCHEMA_NAME, 'wang_xiaoer')){
+  {
     interface initialData{
       asset: number,
       cash_flow: number,
@@ -51,23 +45,20 @@ async function ensureSchema(conn: ChainConnector) {
       }],
     ]);
     for(let [user, data] of customers){
-      await conn.addDocument(ASSET_SCHEMA_NAME, user, JSON.stringify(data));
-      console.log('sample customer %s added', user);
-    }  
+      let exists = await conn.hasDocument(ASSET_SCHEMA_NAME, user)
+      if (!exists){
+        await conn.addDocument(ASSET_SCHEMA_NAME, user, JSON.stringify(data));
+        console.log('sample customer %s added', user);
+      }      
+    }
   }
 }
 
+export const dynamic = 'force-dynamic',
+  revalidate = 0,
+  fetchCache = 'force-no-store';
+
 export default async function RootLayout({ children }) {
-  const version = npmPackage.version;
-  const defaultLang = 'cn';
-  const defaultUser = 'wuming.bank_b';
-
-  const defaultContext: ContextData = {
-    lang: defaultLang,
-    user: defaultUser,
-    version: version,
-  }
-
   let conn = await ChainProvider.connect();
   await ensureSchema(conn);
 
@@ -75,9 +66,7 @@ export default async function RootLayout({ children }) {
     <html>
       <body>
         <Boot />
-        <ContextProvider value={defaultContext}>
-          {children}
-        </ContextProvider>
+        {children}
       </body>
     </html>
   )