+++ components/chain_provider.test.ts
@@ -1,9 +1,8 @@
+import { stringify } from "@bitauth/libauth";
 import ChainProvider from "./chain_provider";
-import { DocumentProperty, PropertyType, ContractDefine, QueryCondition, QueryBuilder } from './chain_sdk';
-
-test('Test Schemas', async () => {
+import { DocumentProperty, PropertyType, ContractDefine, QueryCondition, QueryBuilder, ActorPrivileges } from "@taiyi-io/chain-connector";
+test('schemas', async () => {
     let conn = await ChainProvider.connect();
-    conn.Trace = true;
     const schemaName = 'js-test-case1-schema';
     console.log('schema test begin...');
     {
@@ -66,14 +65,13 @@ test('Test Schemas', async () => {
     console.log('schema interfaces tested');
 });
 
-test('Test Documents', async () => {
+test('documents', async () => {
     const docCount = 10;
     const propertyNameAge = 'age';
     const propertyNameEnabled = 'enabled';
     const schemaName = 'js-test-case2-document';
     const docPrefix = 'js-test-case2-';
     let conn = await ChainProvider.connect();
-    conn.Trace = true;
     console.log('document test begin...');
     {
         let result = await conn.hasSchema(schemaName);
@@ -137,6 +135,9 @@ test('Test Documents', async () => {
         if (totalCount !== records.total) {
             throw new Error('unexpect count ' + records.total + ' => ' + totalCount + ' in case ' + caseName);
         }
+        if (!records.documents){
+            throw new Error("no document available");
+        }
         let recordCount = records.documents.length;
         if (expectResult.length !== recordCount) {
             throw new Error('unexpect result count ' + recordCount + ' => ' + expectResult.length + ' in case ' + caseName);
@@ -187,7 +188,7 @@ test('Test Documents', async () => {
     console.log('document interfaces tested');
 })
 
-test('Test Contracts', async () => {
+test('contracts', async () => {
     const propertyCatalog = "catalog";
     const propertyBalance = "balance";
     const propertyNumber = "number";
@@ -221,7 +222,7 @@ test('Test Contracts', async () => {
         },
     ];
     let conn = await ChainProvider.connect();
-    conn.Trace = true;
+    // conn.Trace = true;
     console.log('contract test begin...');
     {
         let result = await conn.hasSchema(schemaName);
@@ -282,8 +283,19 @@ test('Test Contracts', async () => {
     };
 
     const createContractName = 'contract_create';
+    if (await conn.hasContract(createContractName)) {
+        await conn.withdrawContract(createContractName);
+        console.log('previous contract %s removed', createContractName)
+    }
     await conn.deployContract(createContractName, createContract);
+    expect(await conn.hasContract(createContractName)).toBe(true);
+    let define = await conn.getContract(createContractName);
+    console.log("define:\n%s", JSON.stringify(define));
     const deleteContractName = 'contract_delete';
+    if (await conn.hasContract(deleteContractName)) {
+        await conn.withdrawContract(deleteContractName);
+        console.log('previous contract %s removed', deleteContractName)
+    }
     await conn.deployContract(deleteContractName, deleteContract);
     const docID = 'contract-doc';
     let parameters: string[] = [
@@ -295,14 +307,24 @@ test('Test Contracts', async () => {
         Math.random() > 0.5 ? 'true' : 'false',
         (Math.random() * 200).toFixed(2),
     ];
-    await conn.enableContractTrace(createContractName);
+    let info = await conn.getContractInfo(createContractName);
+    if (!info.trace) {
+        await conn.enableContractTrace(createContractName);
+        console.log('trace enabled');
+    }
+
     await conn.callContract(createContractName, parameters);
     await conn.callContract(deleteContractName, [schemaName, docID]);
     let { total } = await conn.queryContracts(0, 10);
     console.log(total + ' contracts returned before withdraw');
-    await conn.disableContractTrace(createContractName);
+    if (!info.trace) {
+        await conn.disableContractTrace(createContractName);
+        console.log('trace disabled');
+    }
+
     await conn.withdrawContract(createContractName);
     await conn.withdrawContract(deleteContractName);
+
     total = (await conn.queryContracts(0, 10)).total;
     console.log(total + ' contracts returned after withdraw');
 
@@ -312,9 +334,8 @@ test('Test Contracts', async () => {
     console.log('contract interfaces tested');
 })
 
-test('Test Chain', async () => {
+test('chain', async () => {
     let conn = await ChainProvider.connect();
-    conn.Trace = true;
     console.log('chain test begin...');
     let status = await conn.getStatus();
     console.log('world version ' + status.world_version + ', block height ' + status.block_height);
@@ -336,6 +357,9 @@ test('Test Chain', async () => {
         console.log('previous block: => ' + blockData.previous_block);
         console.log('included transactions: ' + blockData.transactions);
         let transactionRecords = await conn.queryTransactions(blockID, 0, maxRecord);
+        if (!transactionRecords.transactions){
+            throw new Error("no transaction available");
+        }
         console.log(transactionRecords.transactions.length + ' / ' + transactionRecords.total + ' transactions returned');
         for (let transID of transactionRecords.transactions) {
             let transactionData = await conn.getTransaction(blockID, transID);
@@ -348,4 +372,118 @@ test('Test Chain', async () => {
     }
 
     console.log('chain interfaces tested');
+})
+
+test('actors', async () => {
+    let conn = await ChainProvider.connect();
+    const schemaName = 'js-test-case5-actors';
+    console.log('actor test begin...');
+
+    {
+        let result = await conn.hasSchema(schemaName);
+        if (result) {
+            console.log('schema ' + schemaName + ' already exsits');
+            await conn.deleteSchema(schemaName);
+            console.log('previous schema ' + schemaName + ' deleted');
+        }
+    }
+    {
+        let properties: DocumentProperty[] = [
+            {
+                name: 'name',
+                type: PropertyType.String
+            },
+            {
+                name: 'age',
+                type: PropertyType.Integer
+            },
+            {
+                name: 'available',
+                type: PropertyType.Boolean
+            }
+        ];
+        await conn.createSchema(schemaName, properties);
+        let current = await conn.getSchema(schemaName);
+        console.log('test schema created ok:\n' + JSON.stringify(current))
+    }
+    let actors = await conn.getSchemaActors(schemaName);
+    console.log(actors);
+
+    const currentGroup = actors[0].group;
+    const actorConfigure: ActorPrivileges[] = [
+        {
+            group: currentGroup,
+            owner: true,
+            executor: true,
+            updater: true,
+            viewer: true,
+        },
+        {
+            group: 'audit',
+            owner: false,
+            executor: false,
+            updater: false,
+            viewer: true,
+        },
+        {
+            group: 'runner',
+            owner: false,
+            executor: true,
+            updater: true,
+            viewer: true,
+        },
+    ];
+    {
+        //test on schema
+        await conn.updateSchemaActors(schemaName, actorConfigure);
+        actors = await conn.getSchemaActors(schemaName);
+        if (JSON.stringify(actors) !== JSON.stringify(actorConfigure)){
+            throw new Error('actors mismatch on schema');
+        }
+        console.log('update schema actors success, current actors: \n%s', stringify(actors, 2));
+    }
+    {        
+        const content = {
+            name: 'hello',
+            age: 20,
+            available: true,
+        }
+        let docID = await conn.addDocument(schemaName, '', JSON.stringify(content));
+        console.log('test doc %s added', docID);
+        await conn.updateDocumentActors(schemaName, docID, actorConfigure);
+        actors = await conn.getDocumentActors(schemaName, docID);
+        if (JSON.stringify(actors) !== JSON.stringify(actorConfigure)){
+            throw new Error('actors mismatch on document');
+        }
+        console.log('update document actors success');
+    }
+    {
+        let define: ContractDefine = {
+            steps: [
+                {
+                    action: "delete_doc",
+                    params: ["@1", "@2"],
+                },
+                {
+                    action: "submit",
+                },
+            ],
+        };
+        const contractName = schemaName;
+        if (await conn.hasContract(contractName)){
+            await conn.withdrawContract(contractName);
+        }
+        await conn.deployContract(contractName, define);
+        console.log('sample contract deployed');
+        await conn.updateContractActors(contractName, actorConfigure);
+        actors = await conn.getContractActors(contractName);
+        if (JSON.stringify(actors) !== JSON.stringify(actorConfigure)){
+            throw new Error('actors mismatch on contract');
+        }
+        console.log('update contract actors success');
+        await conn.withdrawContract(contractName);
+    }
+    
+    await conn.deleteSchema(schemaName);
+    console.log('actor interfaces test success');
 })