+++ CHANGELOG.md
@@ -1,5 +1,32 @@
 # Change Log for Demo B Client
 
+=======
+## [0.4.2] - 2023-03-10
+
+### Changed
+
+- Change output dir of building to 'build' for deploying
+- Change default start port to 3001
+
+## [0.4.1] - 2023-02-26
+
+### Changed
+
+- Allocate a new connector for each connection request
+- Upgrade and adopt to nextjs 13.2
+
+## [0.4.0] - 2023-02-19
+
+### Added
+
+- Allow specifying language directly on the landing page using '/?lang=en'
+- Add chain.project to package.json
+
+### Changed
+
+- Generate pages of different locales on different paths
+- Import @taiyi-io/chain-connector
+
 ## [0.3.1] - 2023-01-26
 
 ### Added